requirements.txt
CacheControl==0.12.5
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
firebase-admin==2.13.0
google-api-core==1.5.1
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-firestore==0.30.0
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.7
lxml==4.2.5
msgpack==0.5.6
mysql-connector==2.1.6
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.3
pytz==2018.7
requests==2.19.1
rsa==4.0
six==1.11.0
transliterate==1.10.2
Unidecode==1.0.22
urllib3==1.23
